MAINT: Fixes for sphinx 7.3 [skip azp] [skip actions] [circle front]
